pyproject.toml
[tool.poetry]
name = "twitter-clone"
version = "0.1.0"
description = ""
authors = ["piegez <jessicamsampaio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.1.3"
djangorestframework = "^3.15.2"
factory-boy = "^3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
